python/requirements.txt
rembg>=2.0.50
Pillow>=10.0.0
imagehash>=4.3.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
